Bugfix error delete message

+++ bot/routers/commands.py
@@ -11,7 +11,8 @@ router = Router()
 
 @router.message(Command('cancel'))
 async def bot_menu_and_start_message(message: Message, state: FSMContext, bot: Bot) -> None:
-    await bot.delete_message(chat_id=message.chat.id, message_id=(await state.get_data()).get('message_id'))
+    if (await state.get_data()).get('message_id'):
+        await bot.delete_message(chat_id=message.chat.id, message_id=(await state.get_data()).get('message_id'))
     
     await message.answer(MENU_MESSAGE, reply_markup=ReplyKeyboardRemove())
     await state.clear()
@@ -19,7 +20,8 @@ async def bot_menu_and_start_message(message: Message, state: FSMContext, bot: B
     
 @router.message(Command('help'))
 async def bot_help(message: Message, state: FSMContext, bot: Bot) -> None:
-    await bot.delete_message(chat_id=message.chat.id, message_id=(await state.get_data()).get('message_id'))
+    if (await state.get_data()).get('message_id'):
+        await bot.delete_message(chat_id=message.chat.id, message_id=(await state.get_data()).get('message_id'))
     
     await  message.answer(HELP_MESSAGE, reply_markup=ReplyKeyboardRemove())
     await state.clear()
